+++ src/site/markdown/release-notes.md
@@ -3,6 +3,7 @@
 ## 2.4.4-SNAPSHOT
 
   * Add optional preamble HTML to login page, defined in lang.json (issue #415)
+  * Add showAsButton to authenticator configuration (in topcat.json) to give an authenticator its own login button (issue #418)
   
 ## 2.4.3 (25th Mar 2019)