+++ src/site/markdown/release-notes.md
@@ -4,6 +4,7 @@
 
   * Add optional preamble HTML to login page, defined in lang.json (issue #415)
   * Add showAsButton to authenticator configuration (in topcat.json) to give an authenticator its own login button (issue #418)
+  * Add extraLoginButtons to facility configuration (in topcat.json) (issue #419)
   
 ## 2.4.3 (25th Mar 2019)