+++ src/site/markdown/release-notes.md
@@ -7,6 +7,7 @@
   * Add extraLoginButtons to facility configuration (in topcat.json) (issue #419)
   * Remove DOI and some login redirections from browser history (issue #424)
   * Allow ids.timeout to be in seconds or minutes as well as milliseconds (issue #349)
+  * Add optional displayName and description to downloadTransportTypes (issue #427)
   
 ## 2.4.3 (25th Mar 2019)